Compilation made cleaner.

+++ 3rdparty/lisp/minilisp.c
@@ -712,7 +712,7 @@ static void gc(void *root) {
     size_t old_nused = mem_nused;
     mem_nused = (size_t)((uint8_t *)scan1 - (uint8_t *)memory);
     if (debug_gc)
-        fprintf(stderr, "GC: %zu bytes out of %zu bytes copied.\n", mem_nused, old_nused);
+	    fprintf(stderr, "GC: %d bytes out of %d bytes copied.\n", (int)mem_nused, (int)old_nused);
     gc_running = false;
 }
